refactor: create Audio component
