+++ src/Components/ChatComponent/components/Call.jsx
@@ -1,0 +1,14 @@
+import { Flex } from "@mantine/core";
+
+export const Call = () => {
+  return (
+    <Flex
+      w="50%"
+      mx="auto"
+      style={{ border: "1px solid red" }}
+      justify="center"
+    >
+      Call
+    </Flex>
+  );
+};
