+++ src/Components/ChatComponent/components/Message/ReceivedMessage.jsx
@@ -1,7 +1,8 @@
 import { Flex, Text, Avatar, DEFAULT_THEME } from "@mantine/core";
 import { DEFAULT_PHOTO, HH_mm } from "../../../../app-constants";
 import { getFullName, parseServerDate, socialMediaIcons } from "../../../utils";
-import { renderContent } from "../../utils";
+import { renderContent, MEDIA_TYPE } from "../../renderContent";
+import { Call } from "../Call";
 
 const { colors } = DEFAULT_THEME;
 
@@ -12,28 +13,33 @@ export const ReceivedMessage = ({ personalInfo, msg }) => {
 
   return (
     <Flex w="100%">
-      <Flex w="90%" direction="column" className="chat-message received">
-        <Flex gap="8">
-          <Avatar src={personalInfo?.photo_url || DEFAULT_PHOTO} />
+      {msg.mtype === MEDIA_TYPE.CALL ? (
+        <Call />
+      ) : (
+        <Flex w="90%" direction="column" className="chat-message received">
+          <Flex gap="8">
+            <Avatar src={personalInfo?.photo_url || DEFAULT_PHOTO} />
 
-          <Flex miw="250px" direction="column" p="8" className="text">
-            <Flex c={colors.gray[7]} align="center" gap="4">
-              <Text size="sm" fw="bold">
-                {getFullName(receivedMsj?.name, receivedMsj?.surname) ||
-                  `#${msg.client_id}`}
+            <Flex miw="250px" direction="column" p="8" className="text">
+              <Flex c={colors.gray[7]} align="center" gap="4">
+                <Text size="sm" fw="bold">
+                  {getFullName(receivedMsj?.name, receivedMsj?.surname) ||
+                    `#${msg.client_id}`}
+                </Text>
+
+                {socialMediaIcons[msg.platform] || null}
+              </Flex>
+
+              {/* {msg.mtype === MEDIA_TYPE.CALL ? <Call /> : renderContent(msg)} */}
+              {renderContent(msg)}
+
+              <Text c={colors.gray[7]} size="sm" ta="end">
+                {parseServerDate(msg.time_sent).format(HH_mm)}
               </Text>
-
-              {socialMediaIcons[msg.platform] || null}
             </Flex>
-
-            {renderContent(msg)}
-
-            <Text c={colors.gray[7]} size="sm" ta="end">
-              {parseServerDate(msg.time_sent).format(HH_mm)}
-            </Text>
           </Flex>
         </Flex>
-      </Flex>
+      )}
     </Flex>
   );
 };
