+++ src/Components/ChatComponent/components/index.js
@@ -1,8 +1,9 @@
-export * from "./ChatListItem"
-export * from "./Media"
-export * from "./Merge"
-export * from "./PersonalData4ClientForm"
-export * from "./ChatInput"
-export * from "./GroupedMessages"
-export * from "./EmojiMessage"
-export * from "./ChatMessages"
+export * from "./ChatListItem";
+export * from "./Media";
+export * from "./Merge";
+export * from "./PersonalData4ClientForm";
+export * from "./ChatInput";
+export * from "./GroupedMessages";
+export * from "./EmojiMessage";
+export * from "./ChatMessages";
+export * from "./Call";
